+++ lib/LibDS/src/Protocols/FRC_2014.cpp
@@ -285,7 +285,7 @@ bool FRC_2014::interpretRobotPacket (const QByteArray& data) {
     /* Update code status & voltage */
     bool hasCode = (integer != 0x37) && (decimal != 0x37);
     config()->setRobotCode (hasCode);
-    config()->updateVoltage (hasCode ? voltage.toFloat() : 0);
+    config()->updateVoltage (hasCode ? voltage.toDouble() : 0);
 
     /* Update simulation status */
     if (receivedRobotPackets() > 10)
